Title: 이차원 배열의 오름차순 정리, Time: 130ms, Memory: 9MB, Status: Passed - CodeTree

+++ 240824/이차원 배열의 오름차순 정리/ascending-order-of-two-dimensional-array.java
@@ -2,7 +2,7 @@ import java.util.*;
 import java.io.*;
 
 public class Main {
-    static int n,k;
+    static long n,k;
     public static void main(String[] args) throws Exception{
         BufferedReader bf = new BufferedReader(new InputStreamReader(System.in));
         n = Integer.parseInt(bf.readLine());
